requirements.txt
# Data Extraction & Processing
pymupdf>=1.24.0
docling==2.69.1
pypdfium2==5.3.0
openai==2.15.0
onnxruntime==1.23.2
python-dotenv==1.0.1
pymysql
chromadb
sentence-transformers
langchain
langchain-community
kiwipiepy
torch
transformers
accelerate
pillow
qwen_vl_utils

# Backend Framework
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
httpx>=0.24.0
python-multipart
pydantic>=2.5.3
pydantic-settings
requests

# Finance & Market Data
yfinance
finance-datareader
pandas
yahoo_fin
requests_html
pykrx
beautifulsoup4
pandas-datareader
lxml_html_clean